lay the groundwork for making open session async
Restructure things a bit so that we can be completely async
about creating sessions.  We want to do this so we can use
callouts to gather session info or to verify info.  Also
make the OpenSession create a parameter list and then create
the session rather than creating the session and changing
attributes.  This is less racy and the session will be complete
and correct before any signals are emitted.