add a way to synchronously run programs on session add/remove/activity_change
This patch adds support for running programs when

 1. A session is added
 2. A session is removed
 3. The activity of a session changes

Executables with the suffix .ck in the directories

 $sysconfdir/ConsoleKit/run-session.d
 (typically /etc/ConsoleKit/run-session.d)

and

 $libdir/ConsoleKit/run-session.d
 (typically /usr/lib/ConsoleKit/run-session.d)

will be run on each event. The former directory is meant to be used
for the system administrator and the latter is meant to be used by
programs.

Only when all programs in these directories have run, ConsoleKit will
resume event processing and e.g. broadcast the event on the system
message bus. Hence, this new mechanism can be used to safely
(e.g. without race conditions) perform operations on certain resources
before programs in the desktop session are told they may use them. The
obvious example here is managing ACL's on /dev such that certain
device nodes are only available to users in local and active sessions.

The environment of the program launched is the environment that the
ConsoleKit daemon was launched with and also the following variables
(variables tagged with [*] may not be set).

 CK_SESSION_ID
 CK_SESSION_TYPE
 CK_SESSION_SEAT_ID
 CK_SESSION_USER_UID
 CK_SESSION_DISPLAY_DEVICE     [*]
 CK_SESSION_X11_DISPLAY_DEVICE [*]
 CK_SESSION_X11_DISPLAY        [*]
 CK_SESSION_REMOTE_HOST_NAME   [*]
 CK_SESSION_IS_ACTIVE
 CK_SESSION_IS_LOCAL

corresponding to the properties of a Session object in question. Each
program is passed exactly one parameter that can assume one of the
following values:

 - session_active_changed:  is_active changed
 - session_added:           the session was added
 - session_removed:         the session was removed

As a safety hatch, there is a timeout of 15 seconds for each program;
if it hasn't exited within 15 seconds, the daemon will send it a
SIGTERM signal, and move on to the next program. The daemon is still
responsive when the program is running - this is to ensure that the
program itself can call into the org.freedesktop.ConsoleKit service.