Get VT from X display if no controlling tty
In some cases a controlling tty won't be able available to
peek at to find the VT of the X display.  In those cases,
we fall back to looking at the XFree86_VT property.

We can do this because we know the display is a local display
since we've already gotten peer credentials from its socket.