get rid of session.d's session_active_changed callout
The 'session_active_changed' callout is an invitaton to racy misuses
since it splits up the session switches into two events. This patch gets
rid of it.

At this point there are two known users of the session.d/ callouts:

  - Suse/Ubuntu ship a script for emulating pam_atconsole style
    /var/run/console management for supporting D-Bus' at_console feature
    properly. Both Martin Pitt and Kay Sievers however are happy with
    having this feature go away. In addition I prepared a patch for
    D-Bus which makes the need for these scripts go away entirely. That
    patch is currently awaiting review.

  - udev's udev-acl tool currently uses this. However this doesn't work
    correctly anyway since it relies on that the CK database is correctly
    dumped before the tool is invoked, which CK does not handle properly
    at this time. In fact fixing this problem is the main reason why I
    came up with this patch series. Kay is also very interested in
    seeing the 'session_active_changed' logic go away.

A later commit introduces a replacement functionality that is less racy.

Please note that this patch does not drop session.d/ in its entirety.
Only the 'session_active_changed' type callout is removed. Scripts that
only rely on 'session_added'/'session_removed' will continue to work
fine, and will continue to be supported.