CKSession: Port to GDBus
Port CKSession to GDBus using gdbus-codegen. The generated
ConsoleKitSession object properties are used where possible.
A test-session binary is generated to test some of the
CkSession properties/DBUS calls. CKSession expects to be
passed a GDBusConnection in ck_session_new but this doesn't
happen because CKManager hasn't been ported yet.
https://github.com/ConsoleKit2/ConsoleKit2/issues/8