Minor fixes
We can't pass around the GVariantBuilder and parse it at the same
time so we'll need to fix the is-local in another way.

Fix the signal emission on the manager.

We may need to use a GDBusObjectManager to get the signals for
interface-added/interface-removed on our Sessions (and eventually
Seats?)