Add a test-seat program
Might as well continue to test everything and print the output to
easily compare changes before and after the port to GDBus. We
don't currently have a way to create seats on the fly to test out
more, but that will come with the multi-seat code when we get there.