CkSession: Fix error handling in dbus_activate
Since we set an initial error message in the event g_signal_emit
isn't handled for the session activate signal, we need to keep
track of that error message. Additionally, that signal call may
return a different error message such as the session is
already activate on the seat so we need to keep track of both since
we'll need to free the memory associated with them.