Add "mode" to the inhibit call
Now we support the mode string of the inhibit dbus call. It can
either be a "block" or "delay". If no delay lock is done, a
system call action will delay 2 seconds to give applications time
to save state. Holding a delay lock will increase the timeout to
a max of 8 seconds.