Add new DBUS method: ListInhibitors
Returns the current list of inhibitors, an array of strings
representing the what, who, why, mode, and two unsigned int32
representing the user ID and process ID.
This is the same method that logind uses. Needed to have inhibit
track the uid/pid of the caller to mirror what logind returns.