ck-remove-directory: return exit code 0 on success
remove_dest_dir() returns the return value of nftw(), which is specified to return 0 on success. The main function was previously returning (ret != TRUE), which would be true in the successful case, as (0 != !0), but true becomes 1 when converted to an int, and a 1 exit code from a process conventionally indicates failure. This commit corrects main's return statement so that a successful run of remove_dest_dir() causes the process to return exit code 0.