ck-remove-directory: remove dest dir as real user
Attempting to remove "<rundir>/user/<uid>" as user <uid> fails because the user typically does not own the "<rundir>/user" directory. This failure causes remove_dest_dir() to call g_error(), which raises a SIGTRAP. This causes a trap diagnostic message to be emitted to the kernel log, which is undesirable. This commit changes remove_dest_dir() so that it removes the contents of the dest dir as user <uid> and then removes the dest dir itself as the real user invoking ck-remove-directory (probably root).