Add support for server managed FDs
These will be needed by session compositors that don't want to
run as root. It should also allow xorg to run as a user as well.

This is what logind exports and session controllers like KDE
expect.

This adds a udev, evdev, libdevattr, and none backend
for device management. More can be added in the future.

This code is based on code in logind, dragonfly BSD's weston
diffs, and libweston; See:
https://freedesktop.org/wiki/Software/systemd/logind/
https://github.com/DragonFlyBSD/DeltaPorts/tree/master/ports/graphics/weston/newport
https://cgit.freedesktop.org/wayland/weston/