Fix ref count warning messages in ck-inhibit
Commit 635648deede5a4 has the side effect of causing a warning
message when the uninhibit signals are sent during the finalize
phase as we can't take a reference there. So we'll need to check
to see if the object was destroyed before sending each emit message.
Additionally, in the manager, move the disconnect signal handler up
since we'll be removing that inhibit anyway.
This should fix:
https://github.com/ConsoleKit2/ConsoleKit2/issues/92
https://github.com/ConsoleKit2/ConsoleKit2/issues/87