Fix make_tmpfs for DragonFly
The previous version of the code assumed that DragonFly
supports passing a string of options to mount(2). This
is not the case and so this lead to a bug where a user's
tmpfs was created with invalid ownership and mode.

We now correctly pass a struct of type tmpfs_mount_info.
The struct is only exported as of DragonFly 5.1 and newer,
so this code will only build on a recent version of DragonFly.