[openbsd] give time for suspend and hibernate to settle
When suspending from a Desktop Environment, the screen lock might not have time
to fire up meaning that on resume there will be a second or so where the content
of the unlocked display can be seen (it cannot be interacted with but still).
Add an arbitrary 2 seconds backgrounded buffer so that the lock has time to
activate.